Custom IP Router
Designed a custom IP router for forwarding of packets in the network.
Used Libpcap library to capture and forward them after modifying the
Ethernet and IP headers. Also added support for ICMP echo and
Traceroute. Used Iperf to evaluate the performance of the packets. Used
RIP protocol as dynamic routing protocol.